requirements.txt
fastapi[all]
pydantic
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
aiofiles
aiosqlite
